refactor: change function name from display to display price
