+++ src/components/lesseeHome-page/PropertyCard.tsx
@@ -3,7 +3,7 @@ import { Card, CardContent } from '@/components/ui/card';
 import Image from 'next/image';
 import { Star } from 'lucide-react';
 import { MiniPropertyI } from '@/types/property';
-import { displayPrice } from '@/function/display';
+import { displayPrice } from '@/function/displayPrice';
 
 export default function PropertyCard({
   image,
